feat(ShareModal): make ShareModal its own component
